Display function keys correctly in status bar menu
Fixes #69

The keyEquivalent property on `NSMenuItem` only displays the first character in the menu. To be able to display key equivalents like “F12”, we need to use font glyphs that render that entire character (which are defined by Apple in a private area of the unicode table: https://developer.apple.com/documentation/appkit/1535851-function-key_unicodes?language=objc)

MASShortcut already supports this with `MASSShortcut.keyCodeStringForKeyEquivalent`. We just need to return this property instead of the more human-readable `keyCodeString` when we want a string that can be used with `NSMenuItem.keyEquivalent`.