Update Main.strings
Hello, I would like to help you with the translation of Rectangle! I translated the missing lines, and also changed "abbreviazioni -> scorciatoie", shortcuts is closer to italian "scorciatoie", "abbreviazioni" is when you write for example Mr. instead of Mister, it's something else. I turned "check for updates automatically" in the shorter "aggiornamenti automatici" because the translated line in the app UI was overlapping with the "check for updates" button, it was too long. I hope you may find this changes useful, thank you for your great work with this app.